# gh-109868: Skip deepcopy memo check for empty memo