@serhiy-storchaka 2023-09-27
The difference is equal to one standard deviation. I am not sure that it is not random.