@eendebakpt 2023-09-27
@serhiy-storchaka Here are some more detailed measurements:
```
import pyperf

setup = """
from copy import deepcopy
d1 = 10
d2 = [1, 2]
d3 = {'hello', 'there'}
d4 = ([1], [2])
"""

runner = pyperf.Runner()
for ii in [1, 2, 3, 4]:
	runner.timeit(name=f"d{ii}", stmt=f"deepcopy(d{ii})", setup=setup)
```
Results (with `--rigorous` option):
```
d1: Mean +- std dev: [main] 275 ns +- 5 ns -> [pr] 226 ns +- 6 ns: 1.22x faster
d2: Mean +- std dev: [main] 1.73 us +- 0.02 us -> [pr] 1.69 us +- 0.04 us: 1.02x faster
d3: Mean +- std dev: [main] 5.98 us +- 0.09 us -> [pr] 5.92 us +- 0.09 us: 1.01x faster
d4: Mean +- std dev: [main] 3.83 us +- 0.07 us -> [pr] 3.81 us +- 0.05 us: 1.01x faster

Geometric mean: 1.06x faster
```
The code change itself is very simple. It avoids a dict lookup, assignment to a local variable `y` and a comparison, which provides a small speedup. For a deepcopy of a larger object the performance improvement will indeed be below the noise.

The performance gain is small, but so is the change and I do not see any downsides to the change.